// apps/condo/admin-ui/address/types.ts
export interface AddressServiceSuggestion<T> {
    value: string;
    unrestricted_value: string;
    data: T;
}

export interface CommonProps<SuggestionType> {
    value?: AddressServiceSuggestion<SuggestionType>;
    url?: string;
    defaultQuery?: string;
    autoload?: boolean;
    delay?: number;
    count?: number;
    onChange?: (suggestion?: AddressServiceSuggestion<SuggestionType>) => void;
    inputProps?: HTMLProps<HTMLInputElement>;
    hintText?: ReactNode;
    renderOption?: (suggestion: AddressServiceSuggestion<SuggestionType>, inputValue: string) => ReactNode;
    containerClassName?: string;
    suggestionsClassName?: string;
    suggestionClassName?: string;
    currentSuggestionClassName?: string;
    hintClassName?: string;
    highlightClassName?: string;
    minChars?: number;
    customInput?: ElementType;
    selectOnBlur?: boolean;
    uid?: string;
    children?: ReactNode | undefined
}